+++ docs/changelog.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- 2025-09-08 3ab11b6 ci: make CLIP import lazy; fix typing; harden workflows
 - 2025-09-08 a0ea7da ci: make CLIP import lazy; fix typing; harden workflows
 - 2025-09-08 a5b08ed Added styles to MKDocs and fixed the workflow scripts
 - 2025-09-08 0a4351c added mkdocs for better documentation